consensus: Move consensus tests into consensus package
